Added search functionality for orders search
